+++ src/App/Dashboard/numLlamadasAtendidas.js
@@ -1,20 +1,24 @@
 import * as React from 'react';
 import { PieChart } from '@mui/x-charts/PieChart';
 
-export default function NumeroLlamadasAtendidas() {
+const data = [
+  { id: 0, value: 10, label: 'Humberto Taboada' },
+  { id: 1, value: 15, label: 'Cristian Nodal' },
+  { id: 2, value: 20, label: 'Adolf Berterame' },
+];
+
+export default function PromedioCalidadServicio() {
   return (
     <PieChart
       series={[
         {
-          data: [
-            { id: 0, value: 10, label: 'Angel Rodriguez' },
-            { id: 1, value: 15, label: 'Esteban Arce' },
-            { id: 2, value: 20, label: 'Rodolfo Cota' },
-          ],
+          data,
+          highlightScope: { faded: 'global', highlighted: 'item' },
+          faded: { innerRadius: 30, additionalRadius: -30, color: 'gray' },
         },
       ]}
-      width={450}
       height={200}
+      width={500}
     />
   );
 }
