+++ src/App/Dashboard/totalProblemasResueltos.js
@@ -4,7 +4,7 @@ import { BarChart } from '@mui/x-charts/BarChart';
 export default function TotalProblemasResueltos() {
   return (
     <BarChart
-      xAxis={[{ scaleType: 'band', data: ['group A', 'group B', 'group C'] }]}
+      xAxis={[{ scaleType: 'band', data: ['Agente 1', 'Agente 2', 'Agente 3']}]}
       series={[{ data: [4, 3, 5] }, { data: [1, 6, 3] }, { data: [2, 5, 6] }]}
       width={500}
       height={300}
